GPM IMERG Datasets

GPM IMERG Datasets are represented by the IMERGDataset AbstractType, which in itself is broken into the IMERGHalfHourly, IMERGDaily and IMERGMonthly Types. For half-hourly and daily datsets, NASA provides not just the output from the final processing run, but also early and late near real-time runs.

The Types that each dataset calls are listed below, along with their function calls.

Type	Early NRT	Late NRT	Final NRT

30 Mins	IMERGHalfHourly	IMERGEarlyHH()	IMERGLateHH()	IMERGFinalHH()

Daily	IMERGDaily	IMERGEarlyDY()	IMERGLateDY()	IMERGFinalDY()

Monthly	IMERGMonthly			IMERGMonthly()

So, for example, if we wanted to get the the Early Near Real-Time IMERG Half-Hourly dataset, we would call the function IMERGEarlyHH, which would return a IMERGHalfHourly data structure (see example at the end of the page).

IMERGDataset Types / Objects

There are three different Types of IMERGDataset:

IMERGHalfHourly, which is used to contain information on half-hourly GPM IMERG datasets

IMERGDaily, which is used to contain information on daily GPM IMERG datasets

IMERGMonthly, which is used to contain information on monthly GPM IMERG datasets

NASAPrecipitation.IMERGHalfHourly — Type

IMERGHalfHourly{ST<:AbstractString, DT<:TimeType} <: IMERGDataset

Object containing information on Half-Hourly IMERG datasets to be downloaded

NASAPrecipitation.IMERGDaily — Type

IMERGDaily{ST<:AbstractString, DT<:TimeType} <: IMERGDataset

Object containing information on Daily IMERG datasets to be downloaded

NASAPrecipitation.IMERGMonthly — Type

IMERGMonthly{ST<:AbstractString, DT<:TimeType} <: IMERGDataset

Object containing information on Montly IMERG datasets to be downloaded

Creating an IMERGHalfHourly dataset

The IMERGHalfHourly dataset structure is used to contain information regarding half-hourly IMERG datasets. There are three functions that create IMERGHalfHourly datasets

IMERGEarlyHH, which is used to retrieve Near Real-Time Early runs

IMERGLateHH, which is used to retrieve Near Real-Time Late runs

IMERGFinalHH, which is used to retrieve the Final post-processing runs

NASAPrecipitation.IMERGEarlyHH — Function

IMERGEarlyHH(
    ST = String,
    DT = Date;
    dtbeg :: TimeType,
    dtend :: TimeType,
    sroot :: AbstractString,
) -> npd :: IMERGHalfHourly{ST,DT}

Creates a IMERGHalfHourly dataset npd to retrieve datasets from the Near Real-Time Early processing runs for Half-Hourly output

Keyword Arguments

dtbeg : Date at which download / analysis of the dataset begins

dtend : Date at which download / analysis of the dataset ends

sroot : The directory in which the folder imergearlyhh will be created for data downloads, storage and analysis

The following fields in npd will be fixed as below:

npdID : imergearlyhh

lname : Early IMERG Half-Hourly

doi : 10.5067/GPM/IMERG/3B-HH-E/06

hroot : https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPML3/GPM3IMERGHHE.06

fpref : 3B-HHR-E.MS.MRG.3IMERG

fsuff : V06B.HDF5

NASAPrecipitation.IMERGLateHH — Function

IMERGLateHH(
    ST = String,
    DT = Date;
    dtbeg :: TimeType,
    dtend :: TimeType,
    sroot :: AbstractString,
) -> npd :: IMERGHalfHourly{ST,DT}

Creates a IMERGHalfHourly dataset npd to retrieve datasets from the Near Real-Time Late processing runs for Half-Hourly output

Keyword Arguments

dtbeg : Date at which download / analysis of the dataset begins

dtend : Date at which download / analysis of the dataset ends

sroot : The directory in which the folder imerglatehh will be created for data downloads, storage and analysis

The following fields in npd will be fixed as below:

npdID : imerglatehh

lname : Late IMERG Half-Hourly

doi : 10.5067/GPM/IMERG/3B-HH-L/06

hroot : https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPML3/GPM3IMERGHHL.06

fpref : 3B-HHR-L.MS.MRG.3IMERG

fsuff : V06B.HDF5

NASAPrecipitation.IMERGFinalHH — Function

IMERGFinalHH(
    ST = String,
    DT = Date;
    dtbeg :: TimeType,
    dtend :: TimeType,
    sroot :: AbstractString,
) -> npd :: IMERGHalfHourly{ST,DT}

Creates a IMERGHalfHourly dataset npd to retrieve datasets from the final post-processing runs for Half-Hourly output

Keyword Arguments

dtbeg : Date at which download / analysis of the dataset begins

dtend : Date at which download / analysis of the dataset ends

sroot : The directory in which the folder imergfinalhh will be created for data downloads, storage and analysis

The following fields in npd will be fixed as below:

npdID : imergfinalhh

lname : Final IMERG Half-Hourly

doi : 10.5067/GPM/IMERG/3B-HH/06

hroot : https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPML3/GPM3IMERGHH.06

fpref : 3B-HHR.MS.MRG.3IMERG

fsuff : V06B.HDF5

Creating an IMERGDaily dataset

The IMERGDaily dataset structure is used to contain information regarding daily IMERG datasets. There are three functions that create IMERGDaily datasets

IMERGEarlyDY, which is used to retrieve Near Real-Time Early runs

IMERGLateDY, which is used to retrieve Near Real-Time Late runs

IMERGFinalDY, which is used to retrieve the Final post-processing runs

NASAPrecipitation.IMERGEarlyDY — Function

IMERGEarlyDY(
    ST = String,
    DT = Date;
    dtbeg :: TimeType,
    dtend :: TimeType,
    sroot :: AbstractString,
) -> npd :: IMERGDaily{ST,DT}

Creates a IMERGDaily dataset npd to retrieve datasets from the Near Real-Time Early processing runs for Daily output

Keyword Arguments

dtbeg : Date at which download / analysis of the dataset begins

dtend : Date at which download / analysis of the dataset ends

sroot : The directory in which the folder imergearlydy will be created for data downloads, storage and analysis

The following fields in npd will be fixed as below:

npdID : imergearlydy

lname : Early IMERG Daily

doi : 10.5067/GPM/IMERGDE/DAY/06

hroot : https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPML3/GPM3IMERGDE.06

fpref : 3B-DAY-E.MS.MRG.3IMERG

fsuff : S000000-E235959.V06.nc4

NASAPrecipitation.IMERGLateDY — Function

IMERGLateDY(
    ST = String,
    DT = Date;
    dtbeg :: TimeType,
    dtend :: TimeType,
    sroot :: AbstractString,
) -> npd :: IMERGDaily{ST,DT}

Creates a IMERGDaily dataset npd to retrieve datasets from the Near Real-Time Late processing runs for Daily output

Keyword Arguments

dtbeg : Date at which download / analysis of the dataset begins

dtend : Date at which download / analysis of the dataset ends

sroot : The directory in which the folder imerglatedy will be created for data downloads, storage and analysis

The following fields in npd will be fixed as below:

npdID : imerglatedy

lname : Late IMERG Daily

doi : 10.5067/GPM/IMERGDL/DAY/06

hroot : https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPML3/GPM3IMERGDL.06

fpref : 3B-DAY-L.MS.MRG.3IMERG

fsuff : S000000-E235959.V06.nc4

NASAPrecipitation.IMERGFinalDY — Function

IMERGFinalDY(
    ST = String,
    DT = Date;
    dtbeg :: TimeType,
    dtend :: TimeType,
    sroot :: AbstractString,
) -> npd :: IMERGDaily{ST,DT}

Creates a IMERGDaily dataset npd to retrieve datasets from the Final post-processing runs for Daily output

Keyword Arguments

dtbeg : Date at which download / analysis of the dataset begins

dtend : Date at which download / analysis of the dataset ends

sroot : The directory in which the folder imergfinaldy will be created for data downloads, storage and analysis

The following fields in npd will be fixed as below:

npdID : imergfinaldy

npdID : Final IMERG Daily

doi : 10.5067/GPM/IMERGDF/DAY/06

hroot : https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPML3/GPM3IMERGDF.06

fpref : 3B-DAY.MS.MRG.3IMERG

fsuff : S000000-E235959.V06.nc4

Creating an IMERGMonthly dataset

The IMERGMonthly dataset structure is used to contain information monthly daily IMERG datasets. There is only one functions that creates IMERGMonthly datasets

IMERGMonthly, which is used to retrieve the Final post-processing runs

NASAPrecipitation.IMERGMonthly — Type

IMERGMonthly(
    ST = String,
    DT = Date;
    dtbeg :: TimeType,
    dtend :: TimeType,
    sroot :: AbstractString,
) -> npd :: IMERGMonthly{ST,DT}

Creates a IMERGMonthly dataset npd to retrieve datasets for Monthly output

Keyword Arguments

dtbeg : Date at which download / analysis of the dataset begins

dtend : Date at which download / analysis of the dataset ends

sroot : The directory in which the folder imergmonthly will be created for data downloads, storage and analysis

The following fields in npd will be fixed as below:

npdID : imergmonthly

lname : IMERG Monthly

doi : 10.5067/GPM/IMERG/3B-MONTH/06

hroot : https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPML3/GPM3IMERGM.06

fpref : 3B-MO.MS.MRG.3IMERG

fsuff : V06B.HDF5

Example

julia> using NASAPrecipitation
julia> npd = IMERGEarlyHH(dtbeg=Date(2017,2,1),dtend=Date(2017,2,1),sroot=pwd())[ Info: 2021-08-08T22:08:10.210 - NASAPrecipitation.jl - Setting up data structure containing information on Early IMERG Half-Hourly data to be downloaded
The NASA Precipitation Dataset {String,Date} has the following properties:
    Dataset ID      (npdID) : imergearlyhh
    Logging Name    (lname) : Early IMERG Half-Hourly
    DOI URL          (doi)  : 10.5067/GPM/IMERG/3B-HH-E/06
    Data Directory  (sroot) : /home/runner/work/NASAPrecipitation.jl/NASAPrecipitation.jl/docs/build/datasets/imergearlyhh
    Date Begin      (dtbeg) : 2017-02-01
    Date End        (dtend) : 2017-02-01
    Timestep                : 30 minutes
    Data Resolution         : 0.1º
    Data Server     (hroot) : https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPM_L3/GPM_3IMERGHHE.06
julia> typeof(npd)NASAPrecipitation.IMERGHalfHourly{String, Date}