NASA Precipitation Datasets

The precipitation datasets that NASAPrecipitation.jl v0.1 currently supports are retrieved from NASA's EOSDIS OPeNDAP servers. They are considered to be a NASAPrecipitationDataset type.

NASAPrecipitationDataset

The NASAPrecipitationDataset AbstractType is meant as an overarching type container for information that identifies various NASA Precipitation Datasets.

NASAPrecipitation.NASAPrecipitationDataset — Type

NASAPrecipitationDataset{ST<:AbstractString, DT<:TimeType}
    npdID :: ST
    lname :: ST
    doi   :: ST
    dtbeg :: DT
    dtend :: DT
    sroot :: ST
    hroot :: ST
    fpref :: ST
    fsuff :: ST
end

Abstract supertype for NASA Precipitation datasets on NASA OPeNDAP Servers.

Fields:

npdID : ID for the NASAPrecipitationDataset, used in determining containing folders and filenames of the NetCDF

lname : The name describing the NASAPrecipitationDataset, used mostly in Logging

doi : The DOI identifier, to be saved into the NetCDF

dtbeg : The start date (Y,M,D) of our download / analysis

dtend : The end date (Y,M,D) of our download / analysis

sroot : The directory in which to save our downloads and analysis files to

hroot : The URL of the NASA's EOSDIS OPeNDAP server for which this dataset is stored

fpref : The prefix component of the NetCDF files to be downloaded

fsuff : The suffix component of the NetCDF files to be downloaded

Of these fields, only dtbeg, dtend and sroot are user-defined. All other fields are predetermined depending on the type of NASA Precipitation Dataset called.

All subTypes of the NASAPrecipitationDataset have the same basic fields and structure.

julia> using NASAPrecipitation

julia> npd = IMERGFinalHH(dtbeg=Date(2017,2,1),dtend=Date(2017,2,1),sroot=pwd())
[ Info: 2021-08-07T05:29:14.913 - NASAPrecipitation.jl - Setting up data structure containing information on Final IMERG Half-Hourly data to be downloaded
The NASA Precipitation Dataset {String,Date} has the following properties:
    Dataset ID      (npdID) : imergfinalhh
    Logging Name    (lname) : Final IMERG Half-Hourly
    DOI URL          (doi)  : 10.5067/GPM/IMERG/3B-HH/06
    Data Directory  (sroot) : /home/runner/work/NASAPrecipitation.jl/NASAPrecipitation.jl/docs/build/datasets/imergfinalhh
    Date Begin      (dtbeg) : 2017-02-01
    Date End        (dtend) : 2017-02-01
    Timestep                : 30 minutes
    Data Resolution         : 0.1º
    Data Server     (hroot) : https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPM_L3/GPM_3IMERGHH.06

julia> npd.dtbeg
2017-02-01

julia> npd.sroot
"/home/runner/work/NASAPrecipitation.jl/NASAPrecipitation.jl/docs/build/datasets/imergfinalhh"

julia> npd.fpref
"3B-HHR.MS.MRG.3IMERG"

julia> typeof(npd)
NASAPrecipitation.IMERGHalfHourly{String, Date}

julia> typeof(npd) <: NASAPrecipitation.NASAPrecipitationDataset
true

Supported NASA Datasets

The following datasets are supported in v0.1:

GPM IMERGv6, 0.1º resolution, 60ºS to 60ºN, represented by the IMERGDataset Type

Final runs for Half-Hourly, Daily and Monthly Data

Early and Late runs of Half-Hourly and Daily Data

TRMM TMPAv7, 0.25º resolution, 50ºS to 50ºN, represented by the TRMMataset Type

Final runs for 3-Hourly, Daily and Monthly Data

Near Real-Time runs of 3-Hourly and Daily Data

NASAPrecipitation.IMERGDataset — Type

IMERGDataset <: NASAPrecipitationDataset

Abstract supertype for GPM IMERG datasets on NASA OPeNDAP Servers, a subType of NASAPrecipitationDataset.

NASAPrecipitation.TRMMDataset — Type

TRMMDataset <: NASAPrecipitationDataset

Abstract supertype for TRMM TMPA datasets on NASA OPeNDAP Servers, a subType of NASAPrecipitationDataset.

Each of these Abstract superTypes has its own set of Types. See the respective pages for GPM IMERG and TRMM TMPA for more information.