Download API and Examples

Base.download — Function

download(
    npd :: NASAPrecipitationDataset,
    geo :: GeoRegion
) -> nothing

Downloads a NASA Precipitation dataset specified by npd for a GeoRegion specified by geo

Arguments

npd : a NASAPrecipitationDataset specifying the dataset details and date download range

geo : a GeoRegion (see GeoRegions.jl) that sets the geographic bounds of the data array in lon-lat