Dummy Datasets

Dummy datasets are meant to specify the paths to the NASAPrecipitation root directory, and LandSea Dataset directory, without needing inputs for date.

NASAPrecipitation.IMERGDummy — Type

IMERGDummy{ST<:AbstractString} <: IMERGDataset

Object containing information on path information for IMERG datasets, with the following fields:

rootpath : path of folder containing all GPM IMERG datasets

maskpath : path of folder containing IMERG LandSea mask data

NASAPrecipitation.TRMMDummy — Type

TRMMDummy{ST<:AbstractString} <: TRMMDataset

Object containing information on path information for TRMM datasets, with the following fields:

rootpath : path of folder containing all GPM IMERG datasets

maskpath : path of folder containing IMERG LandSea mask data

julia> using NASAPrecipitation
julia> npd = IMERGDummy(path=homedir())[ Info: 2023-05-28T17:13:55.854 - NASAPrecipitation.jl - Setting up data structure containing path information for IMERG Dummy Dataset
The IMERG Dummy Dataset {String} has the following properties:
    Root Directory (rootpath) : /home/runner
    Mask Directory (maskpath) : /home/runner/imergmask
    Data Resolution           : 0.1º
julia> npd = TRMMDummy(path=homedir())[ Info: 2023-05-28T17:13:55.865 - NASAPrecipitation.jl - Setting up data structure containing path information for TRMM Dummy Dataset
The TRMM Dummy Dataset {String} has the following properties:
    Root Directory (rootpath) : /home/runner
    Mask Directory (maskpath) : /home/runner/trmmmask
    Data Resolution           : 0.25º
julia> isdir(npd.maskpath)true

API

NASAPrecipitation.IMERGDummy — Type

IMERGDummy(
    ST = String;
    path  :: AbstractString = homedir(),
) -> npd :: IMERGDummy

Creates a IMERGDummy dataset npd that contains information on the data and mask paths

Keyword Arguments

path : The directory in which the folder imergmask will be created for the IMERG Land-Sea mask

NASAPrecipitation.TRMMDummy — Type

TRMMDummy(
    ST = String;
    path  :: AbstractString = homedir(),
) -> npd :: IMERGDummy

Creates a TRMMDummy dataset npd that contains information on the data and mask paths

Keyword Arguments

path : The directory in which the folder trmmmask will be created for the TRMM Land-Sea mask