GPM IMERG Datasets

GPM IMERG Datasets are represented by the IMERGDataset AbstractType, which in itself is broken into the IMERGHalfHourly, IMERGDaily and IMERGMonthly Types. For half-hourly and daily datsets, NASA provides not just the output from the final processing run, but also early and late near real-time runs.

The Types that each dataset calls are listed below, along with their function calls.

Type	Early NRT	Late NRT	Final NRT

30 Mins	IMERGHalfHourly	IMERGEarlyHH()	IMERGLateHH()	IMERGFinalHH()

Daily	IMERGDaily	IMERGEarlyDY()	IMERGLateDY()	IMERGFinalDY()

Monthly	IMERGMonthly			IMERGMonthly()

So, for example, if we wanted to get the the Early Near Real-Time IMERG Half-Hourly dataset, we would call the function IMERGEarlyHH, which would return a IMERGHalfHourly data structure (see example at the end of the page).

Setup

using NASAPrecipitation

IMERGDataset Types / Objects

There are three different Types of IMERGDataset:

IMERGHalfHourly, which is used to contain information on half-hourly GPM IMERG datasets

IMERGDaily, which is used to contain information on daily GPM IMERG datasets

IMERGMonthly, which is used to contain information on monthly GPM IMERG datasets

NASAPrecipitation.IMERGHalfHourly — Type

IMERGHalfHourly{ST<:AbstractString, DT<:TimeType} <: IMERGDataset

Object containing information on Half-Hourly IMERG datasets to be downloaded

NASAPrecipitation.IMERGDaily — Type

IMERGDaily{ST<:AbstractString, DT<:TimeType} <: IMERGDataset

Object containing information on Daily IMERG datasets to be downloaded

NASAPrecipitation.IMERGMonthly — Type

IMERGMonthly{ST<:AbstractString, DT<:TimeType} <: IMERGDataset

Object containing information on Montly IMERG datasets to be downloaded

Creating an IMERGHalfHourly dataset

The IMERGHalfHourly dataset structure is used to contain information regarding half-hourly IMERG datasets. There are three functions that create IMERGHalfHourly datasets

IMERGEarlyHH, which is used to retrieve Near Real-Time Early runs

IMERGLateHH, which is used to retrieve Near Real-Time Late runs

IMERGFinalHH, which is used to retrieve the Final post-processing runs

npd = IMERGEarlyHH(start=Date(2017,2,1),stop=Date(2017,2,1))

The NASA Precipitation Dataset {String,Date} has the following properties:
    Dataset ID            (ID) : imergearlyhh
    Logging Name        (name) : Early IMERG Half-Hourly
    DOI URL              (doi) : 10.5067/GPM/IMERG/3B-HH-E/06
    Data Directory  (datapath) : /home/runner/imergearlyhh
    Mask Directory  (maskpath) : /home/runner/imergmask
    Date Begin         (start) : 2017-02-01
    Date End            (stop) : 2017-02-01
    Timestep                   : 30 minutes
    Data Resolution            : 0.1º
    Data Server        (hroot) : https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPM_L3/GPM_3IMERGHHE.06

typeof(npd)

NASAPrecipitation.IMERGHalfHourly{String, Date}

npd = IMERGLateHH(start=Date(2017,2,1),stop=Date(2017,2,1))

The NASA Precipitation Dataset {String,Date} has the following properties:
    Dataset ID            (ID) : imerglatehh
    Logging Name        (name) : Late IMERG Half-Hourly
    DOI URL              (doi) : 10.5067/GPM/IMERG/3B-HH-L/06
    Data Directory  (datapath) : /home/runner/imerglatehh
    Mask Directory  (maskpath) : /home/runner/imergmask
    Date Begin         (start) : 2017-02-01
    Date End            (stop) : 2017-02-01
    Timestep                   : 30 minutes
    Data Resolution            : 0.1º
    Data Server        (hroot) : https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPM_L3/GPM_3IMERGHHL.06

typeof(npd)

NASAPrecipitation.IMERGHalfHourly{String, Date}

We see as above that whether a dataset is EarlyNRT or LateNRT or Final doesn't matter, it will return the same dataset type. What changes will be the values in the fields, not the dataset structure or type itself.

Creating an IMERGDaily dataset

The IMERGDaily dataset structure is used to contain information regarding daily IMERG datasets. There are three functions that create IMERGDaily datasets

IMERGEarlyDY, which is used to retrieve Near Real-Time Early runs

IMERGLateDY, which is used to retrieve Near Real-Time Late runs

IMERGFinalDY, which is used to retrieve the Final post-processing runs

npd = IMERGFinalDY(start=Date(2017,2,5),stop=Date(2017,2,5))

The NASA Precipitation Dataset {String,Date} has the following properties:
    Dataset ID           (ID) : imergv7finaldy
    Logging Name       (name) : Final IMERGv7 Daily
    DOI URL             (doi) : 10.5067/GPM/IMERGDF/DAY/07
    Data Directory (datapath) : /home/runner/imergv7finaldy
    Mask Directory (maskpath) : /home/runner/imergmask
    Date Begin        (start) : 2017-02-01
    Date End           (stop) : 2017-02-28
    Timestep                  : 1 Day
    Data Resolution           : 0.1º
    Data Server       (hroot) : https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPM_L3/GPM_3IMERGDF.07

typeof(npd)

NASAPrecipitation.IMERGDaily{String, Date}

Info

Notice here, that npd.start and npd.stop define the whole month of Feb 2017, as, for efficiency purposes, IMERGDaily datasets are designed to select data by entire months.

Creating an IMERGMonthly dataset

The IMERGMonthly dataset structure is used to contain information monthly daily IMERG datasets. There is only one functions that creates IMERGMonthly datasets

IMERGMonthly, which is used to retrieve the Final post-processing runs

npd = IMERGMonthly(start=Date(2017,6,1),stop=Date(2017,8,15))

The NASA Precipitation Dataset {String,Date} has the following properties:
    Dataset ID            (ID) : imergv7monthly
    Logging Name        (name) : IMERGv7 Monthly
    DOI URL              (doi) : 10.5067/GPM/IMERG/3B-MONTH/07
    Data Directory  (datapath) : /home/runner/imergv7monthly
    Mask Directory  (maskpath) : /home/runner/imergmask
    Date Begin         (start) : 2017-01-01
    Date End            (stop) : 2017-12-31
    Timestep                   : 1 Month
    Data Resolution            : 0.1º
    Data Server        (hroot) : https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPM_L3/GPM_3IMERGM.07

typeof(npd)

IMERGMonthly{String, Date}

Info

Notice here, that npd.start and npd.stop define the whole year of 2017, as, for efficiency purposes, IMERGMonthly datasets are designed to select data by entire years.

API

NASAPrecipitation.IMERGEarlyHH — Function

IMERGEarlyHH(
    ST = String,
    DT = Date;
    start :: TimeType,
    stop  :: TimeType,
    path  :: AbstractString = homedir(),
) -> npd :: IMERGHalfHourly{ST,DT}

Creates a IMERGHalfHourly dataset npd to retrieve datasets from the Near Real-Time Early processing runs for Half-Hourly output

Keyword Arguments

start : Date at which download / analysis of the dataset begins

stop : Date at which download / analysis of the dataset ends

path : The directory in which the folder imergearlyhh will be created for data downloads, storage and analysis, default is the home directoy called by homedir()

The following fields in npd will be fixed as below:

ID : imergearlyhh

name : Early IMERG Half-Hourly

doi : 10.5067/GPM/IMERG/3B-HH-E/06

hroot : https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPML3/GPM3IMERGHHE.06

fpref : 3B-HHR-E.MS.MRG.3IMERG

fsuff : V06B.HDF5

NASAPrecipitation.IMERGLateHH — Function

IMERGLateHH(
    ST = String,
    DT = Date;
    start :: TimeType,
    stop :: TimeType,
    path :: AbstractString = homedir(),
) -> npd :: IMERGHalfHourly{ST,DT}

Creates a IMERGHalfHourly dataset npd to retrieve datasets from the Near Real-Time Late processing runs for Half-Hourly output

Keyword Arguments

start : Date at which download / analysis of the dataset begins

stop : Date at which download / analysis of the dataset ends

path : The directory in which the folder imerglatehh will be created for data downloads, storage and analysis, default is the home directoy called by homedir()

The following fields in npd will be fixed as below:

ID : imerglatehh

name : Late IMERG Half-Hourly

doi : 10.5067/GPM/IMERG/3B-HH-L/06

hroot : https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPML3/GPM3IMERGHHL.06

fpref : 3B-HHR-L.MS.MRG.3IMERG

fsuff : V06B.HDF5

NASAPrecipitation.IMERGFinalHH — Function

IMERGFinalHH(
    ST = String,
    DT = Date;
    start :: TimeType,
    stop :: TimeType,
    path :: AbstractString = homedir(),
) -> npd :: IMERGHalfHourly{ST,DT}

Creates a IMERGHalfHourly dataset npd to retrieve datasets from the final post-processing runs for Half-Hourly output

Keyword Arguments

start : Date at which download / analysis of the dataset begins

stop : Date at which download / analysis of the dataset ends

path : The directory in which the folder imergfinalhh will be created for data downloads, storage and analysis, default is the home directoy called by homedir()

The following fields in npd will be fixed as below:

ID : imergfinalhh

name : Final IMERG Half-Hourly

doi : 10.5067/GPM/IMERG/3B-HH/06

hroot : https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPML3/GPM3IMERGHH.06

fpref : 3B-HHR.MS.MRG.3IMERG

fsuff : V06B.HDF5

NASAPrecipitation.IMERGEarlyDY — Function

IMERGEarlyDY(
    ST = String,
    DT = Date;
    start :: TimeType,
    stop  :: TimeType,
    path :: AbstractString = homedir(),
) -> npd :: IMERGDaily{ST,DT}

Creates a IMERGDaily dataset npd to retrieve datasets from the Near Real-Time Early processing runs for Daily output

Keyword Arguments

start : Date at which download / analysis of the dataset begins

stop : Date at which download / analysis of the dataset ends

path : The directory in which the folder imergearlydy will be created for data downloads, storage and analysis, default is the home directoy called by homedir()

The following fields in npd will be fixed as below:

ID : imergearlydy

name : Early IMERG Daily

doi : 10.5067/GPM/IMERGDE/DAY/06

hroot : https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPML3/GPM3IMERGDE.06

fpref : 3B-DAY-E.MS.MRG.3IMERG

fsuff : S000000-E235959.V06.nc4

NASAPrecipitation.IMERGLateDY — Function

IMERGLateDY(
    ST = String,
    DT = Date;
    start :: TimeType,
    stop  :: TimeType,
    path :: AbstractString = homedir(),
) -> npd :: IMERGDaily{ST,DT}

Creates a IMERGDaily dataset npd to retrieve datasets from the Near Real-Time Late processing runs for Daily output

Keyword Arguments

start : Date at which download / analysis of the dataset begins

stop : Date at which download / analysis of the dataset ends

path : The directory in which the folder imerglatedy will be created for data downloads, storage and analysis, default is the home directoy called by homedir()

The following fields in npd will be fixed as below:

ID : imerglatedy

name : Late IMERG Daily

doi : 10.5067/GPM/IMERGDL/DAY/06

hroot : https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPML3/GPM3IMERGDL.06

fpref : 3B-DAY-L.MS.MRG.3IMERG

fsuff : S000000-E235959.V06.nc4

NASAPrecipitation.IMERGFinalDY — Function

IMERGFinalDY(
    ST = String,
    DT = Date;
    start :: TimeType,
    stop  :: TimeType,
    path :: AbstractString = homedir(),
) -> npd :: IMERGDaily{ST,DT}

Creates a IMERGDaily dataset npd to retrieve datasets from the Final post-processing runs for Daily output

Keyword Arguments

start : Date at which download / analysis of the dataset begins

stop : Date at which download / analysis of the dataset ends

path : The directory in which the folder imergfinaldy will be created for data downloads, storage and analysis, default is the home directoy called by homedir()

The following fields in npd will be fixed as below:

ID : imergfinaldy

ID : Final IMERG Daily

doi : 10.5067/GPM/IMERGDF/DAY/06

hroot : https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPML3/GPM3IMERGDF.06

fpref : 3B-DAY.MS.MRG.3IMERG

fsuff : S000000-E235959.V06.nc4

NASAPrecipitation.IMERGMonthly — Type

IMERGMonthly(
    ST = String,
    DT = Date;
    start :: TimeType,
    stop  :: TimeType,
    path  :: AbstractString = homedir(),
) -> npd :: IMERGMonthly{ST,DT}

Creates a IMERGMonthly dataset npd to retrieve datasets for Monthly output

Keyword Arguments

start : Date at which download / analysis of the dataset begins

stop : Date at which download / analysis of the dataset ends

path : The directory in which the folder imergmonthly will be created for data downloads, storage and analysis, default is the home directoy called by homedir()

The following fields in npd will be fixed as below:

ID : imergmonthly

name : IMERG Monthly

doi : 10.5067/GPM/IMERG/3B-MONTH/06

hroot : https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPML3/GPM3IMERGM.06

fpref : 3B-MO.MS.MRG.3IMERG

fsuff : V06B.HDF5