An Overview of the NASA Precipitation Datasets

A NASA Precipitation Dataset is defined to be a dataset from the NASA Global Precipitation Measurement (GPM) Mission. As of now, this package supports the following Level 3 products:

GPM IMERGv6, 0.1º resolution, 60ºS to 60ºN

Final runs for Half-Hourly, Daily and Monthly Data

Early and Late runs of Half-Hourly and Daily Data

TRMM TMPAv7, 0.25º resolution, 50ºS to 50ºN

Final runs for 3-Hourly, Daily and Monthly Data

Near Real-Time runs of 3-Hourly and Daily Data

The NASAPrecipitationDataset AbstractType is meant as an overarching type container for information that identifies various NASA Precipitation Datasets. The NASAPrecipitationDataset Type has two sub-Types:

The IMERGDataset for the GPM IMERG Dataset

The TRMMDataset for the TRMM TMPA Dataset

Each of these Abstract superTypes has its own set of Types. See the summary table below for an overview, and the respective pages for GPM IMERG and TRMM TMPA for more information.

NASAPrecipitation.NASAPrecipitationDataset — Type

NASAPrecipitationDataset{ST<:AbstractString, DT<:TimeType}
    ID    :: ST
    name  :: ST
    doi   :: ST
    start :: DT
    stop  :: DT
    datapath :: ST
    maskpath :: ST
    hroot :: ST
    fpref :: ST
    fsuff :: ST
end

Abstract supertype for NASA Precipitation datasets on NASA OPeNDAP Servers.

Fields:

ID : ID for the NASAPrecipitationDataset, used in determining containing folders and filenames of the NetCDF

name : The name describing the NASAPrecipitationDataset, used mostly in Logging

doi : The DOI identifier, to be saved into the NetCDF

start : The start date (Y,M,D) of our download / analysis

stop : The end date (Y,M,D) of our download / analysis

datapath : The directory in which to save our downloads and analysis files to

maskpath : The directory in which to save our corresponding landsea mask datasets

hroot : The URL of the NASA's EOSDIS OPeNDAP server for which this dataset is stored

fpref : The prefix component of the NetCDF files to be downloaded

fsuff : The suffix component of the NetCDF files to be downloaded

Of these fields, only start, stop, datapath and maskpath are user-defined. All other fields are predetermined depending on the type of NASA Precipitation Dataset called.

NASAPrecipitation.IMERGDataset — Type

IMERGDataset <: NASAPrecipitationDataset

Abstract supertype for GPM IMERG datasets on NASA OPeNDAP Servers, a subType of NASAPrecipitationDataset.

NASAPrecipitation.TRMMDataset — Type

TRMMDataset <: NASAPrecipitationDataset

Abstract supertype for TRMM TMPA datasets on NASA OPeNDAP Servers, a subType of NASAPrecipitationDataset.

All subTypes of the NASAPrecipitationDataset have the same basic fields and structure.

Summary Table

SuperType	Type	Early NRT	(Late) NRT	Final NRT

IMERG 30 Mins	IMERGDataset	IMERGHalfHourly	IMERGEarlyHH()	IMERGLateHH()	IMERGFinalHH()

IMERG Daily	IMERGDataset	IMERGDaily	IMERGEarlyDY()	IMERGLateDY()	IMERGFinalDY()

IMERG Monthly	IMERGDataset	IMERGMonthly			IMERGMonthly()

TRMM 3 Hour	TRMMDataset	TRMM3Hourly		TRMM3HourlyNRT()	TRMM3Hourly()

TRMM Daily	TRMMDataset	TRMMDaily		TRMMDailyNRT()	TRMMDaily()

TRMM Monthly	TRMMDataset	TRMMMonthly			TRMMMonthly()