TRMM TMPA Datasets

TRMM TMPA Datasets are represented by the TRMMDataset AbstractType, which in itself is broken into the TRMM3Hourly, TRMMDaily and TRMMMonthly Types. For 3-hourly and daily datsets, NASA provides not just the output from the final processing run, but also near real-time runs.

The Types that each dataset calls are listed below, along with their function calls.

Type	Near Real-Time	Final

3 Hour	TRMM3Hourly	TRMM3HourlyNRT()	TRMM3Hourly()

Daily	TRMMDaily	TRMMDailyNRT()	TRMMDaily()

Monthly	TRMMMonthly		TRMMMonthly()

So, for example, if we wanted to get the the Near Real-Time TRMM 3-Hourly dataset, we would call the function TRMMMonthly(), which would return a TRMMMonthly data structure (see example at the end of the page).

Setup

using NASAPrecipitation

TRMMDataset Types / Objects

There are three different Types of TRMMDataset:

TRMM3Hourly, which is used to contain information on half-hourly TRMM TMPA datasets

TRMMDaily, which is used to contain information on daily TRMM TMPA datasets

TRMMMonthly, which is used to contain information on monthly TRMM TMPA datasets

NASAPrecipitation.TRMM3Hourly — Type

TRMM3Hourly{ST<:AbstractString, DT<:TimeType} <: TRMMDataset

Object containing information on 3-Hourly TRMM datasets to be downloaded

NASAPrecipitation.TRMMDaily — Type

TRMMDaily{ST<:AbstractString, DT<:TimeType} <: TRMMDataset

Object containing information on Daily TRMM datasets to be downloaded

NASAPrecipitation.TRMMMonthly — Type

TRMMMonthly{ST<:AbstractString, DT<:TimeType} <: TRMMDataset

Object containing information on Monthly TRMM datasets to be downloaded

Creating a TRMM3Hourly dataset

The TRMM3Hourly dataset structure is used to contain information regarding 3-hourly TRMM datasets. There are two functions that create TRMM3Hourly datasets

TRMM3HourlyNRT, which is used to retrieve Near Real-Time runs

TRMM3Hourly, which is used to retrieve the Final post-processing runs

npd = TRMM3HourlyNRT(start=Date(2017,2,1),stop=Date(2017,2,1))

The NASA Precipitation Dataset {String,Date} has the following properties:
    Dataset ID         (npdID) : trmm3hourlynrt
    Logging Name       (lname) : Near Real-Time TRMM 3-Hourly
    DOI URL              (doi) : 10.5067/TRMM/TMPA/3H-E/7
    Data Directory  (datapath) : /home/runner/trmm3hourlynrt
    Mask Directory  (maskpath) : /home/runner/trmmmask
    Date Begin         (start) : 2017-02-01
    Date End            (stop) : 2017-02-28
    Timestep                   : 3 Hourly
    Data Resolution            : 0.25º
    Data Server        (hroot) : https://disc2.gesdisc.eosdis.nasa.gov/opendap/TRMM_RT/TRMM_3B42RT.7

typeof(npd)

TRMM3Hourly{String, Date}

npd = TRMM3Hourly(start=Date(2017,2,1),stop=Date(2017,2,1))

The NASA Precipitation Dataset {String,Date} has the following properties:
    Dataset ID         (npdID) : trmm3hourly
    Logging Name       (lname) : Final TRMM 3-Hourly
    DOI URL              (doi) : 10.5067/TRMM/TMPA/3H/7
    Data Directory  (datapath) : /home/runner/trmm3hourly
    Mask Directory  (maskpath) : /home/runner/trmmmask
    Date Begin         (start) : 2017-02-01
    Date End            (stop) : 2017-02-28
    Timestep                   : 3 Hourly
    Data Resolution            : 0.25º
    Data Server        (hroot) : https://disc2.gesdisc.eosdis.nasa.gov/opendap/TRMM_L3/TRMM_3B42.7

typeof(npd)

TRMM3Hourly{String, Date}

We see as above that whether a dataset is EarlyNRT or LateNRT or Final doesn't matter, it will return the same dataset type. What changes will be the values in the fields, not the dataset structure or type itself.

Creating a TRMMDaily dataset

The TRMMDaily dataset structure is used to contain information regarding Daily TRMM datasets. There are two functions that create TRMMDaily datasets

TRMMDailyNRT, which is used to retrieve Near Real-Time runs

TRMMDaily, which is used to retrieve the Final post-processing runs

npd = TRMMDailyNRT(start=Date(2017,2,5),stop=Date(2017,2,5))

The NASA Precipitation Dataset {String,Date} has the following properties:
    Dataset ID         (npdID) : trmmdailynrt
    Logging Name        (lname): Near Real-Time TRMM Daily
    DOI URL              (doi) : 10.5067/TRMM/TMPA/DAY-E/7
    Data Directory  (datapath) : /home/runner/trmmdailynrt
    Mask Directory  (maskpath) : /home/runner/trmmmask
    Date Begin         (start) : 2017-02-01
    Date End            (stop) : 2017-02-28
    Timestep                   : 1 Day
    Data Resolution            : 0.25º
    Data Server        (hroot) : https://disc2.gesdisc.eosdis.nasa.gov/opendap/TRMM_RT/TRMM_3B42RT_Daily.7

typeof(npd)

TRMMDaily{String, Date}

Info

Notice here, that npd.start and npd.stop define the whole month of Feb 2017, as, for efficiency purposes, TRMMDaily datasets are designed to select data by entire months.

Creating a TRMMMonthly dataset

The TRMMMonthly dataset structure is used to contain information regarding Monthly TRMM datasets. The function used to create TRMMMonthly datasets is

TRMMMonthly, which is used to retrieve the Final post-processing runs

npd = TRMMMonthly(start=Date(2017,6,1),stop=Date(2017,8,15))

The NASA Precipitation Dataset {String,Date} has the following properties:
    Dataset ID         (npdID) : trmmmonthly
    Logging Name       (lname) : TRMM Monthly (TMPA 3B43)
    DOI URL              (doi) : 10.5067/TRMM/TMPA/MONTH/7
    Data Directory  (datapath) : /home/runner/trmmmonthly
    Mask Directory  (maskpath) : /home/runner/trmmmask
    Date Begin         (start) : 2017-01-01
    Date End            (stop) : 2017-12-31
    Timestep                   : 1 Month
    Data Resolution            : 0.25º
    Data Server        (hroot) : https://disc2.gesdisc.eosdis.nasa.gov/opendap/TRMM_L3/TRMM_3B43.7

typeof(npd)

TRMMMonthly{String, Date}

Info

Notice here, that npd.start and npd.stop define the whole year of 2017, as, for efficiency purposes, TRMMMonthly datasets are designed to select data by entire years.

API

NASAPrecipitation.TRMM3HourlyNRT — Function

TRMM3HourlyNRT(
    ST = String,
    DT = Date;
    start :: TimeType,
    stop  :: TimeType,
    path  :: AbstractString = homedir(),
) -> npd :: TRMM3Hourly{ST,DT}

Creates a TRMM3Hourly dataset npd to retrieve datasets from the Near Real-Time runs for 3-Hourly output

Keyword Arguments

start : Date at which download / analysis of the dataset begins

stop : Date at which download / analysis of the dataset ends

path : The directory in which the folder trmm3hourlynrt will be created for data downloads, storage and analysis, default is the home directoy called by homedir()

The following fields in npd will be fixed as below:

npdID : trmm3hourlynrt

lname : Near Real-Time TRMM 3-Hourly

doi : 10.5067/TRMM/TMPA/3H-E/7

hroot : https://disc2.gesdisc.eosdis.nasa.gov/opendap/TRMMRT/TRMM3B42RT.7

fpref : 3B42RT

fsuff : 7R2.nc4

NASAPrecipitation.TRMM3Hourly — Type

TRMM3Hourly(
    ST = String,
    DT = Date;
    start :: TimeType,
    stop  :: TimeType,
    path  :: AbstractString = homedir(),
) -> npd :: TRMM3Hourly{ST,DT}

Creates a TRMM3Hourly dataset npd to retrieve datasets from the Final post-processing runs for 3-Hourly output

Keyword Arguments

start : Date at which download / analysis of the dataset begins

stop : Date at which download / analysis of the dataset ends

path : The directory in which the folder trmm3hourly will be created for data downloads, storage and analysis, default is the home directoy called by homedir()

The following fields in npd will be fixed as below:

npdID : trmm3hourly

lname : Final TRMM 3-Hourly

doi : 10.5067/TRMM/TMPA/3H/7

hroot : https://disc2.gesdisc.eosdis.nasa.gov/opendap/TRMML3/TRMM3B42.7

fpref : 3B42

fsuff : 7.HDF

NASAPrecipitation.TRMMDailyNRT — Function

TRMMDailyNRT(
    ST = String,
    DT = Date;
    start :: TimeType,
    stop  :: TimeType,
    path  :: AbstractString = homedir(),
) -> npd :: TRMMDaily{ST,DT}

Creates a TRMMDaily dataset npd to retrieve datasets from the Near Real-Time processing runs for Daily output

Keyword Arguments

start : Date at which download / analysis of the dataset begins

stop : Date at which download / analysis of the dataset ends

path : The directory in which the folder trmmdailynrt will be created for data downloads, storage and analysis, default is the home directoy called by homedir()

The following fields in npd will be fixed as below:

npdID : trmmdailynrt

lname : Near Real-Time TRMM Daily

doi : 10.5067/TRMM/TMPA/DAY-E/7

hroot : https://disc2.gesdisc.eosdis.nasa.gov/opendap/TRMMRT/TRMM3B42RT_Daily.7

fpref : 3B42RT_Daily

fsuff : 7.nc4

NASAPrecipitation.TRMMDaily — Type

TRMMDaily(
    ST = String,
    DT = Date;
    start :: TimeType,
    stop  :: TimeType,
    path  :: AbstractString = homedir(),
) -> npd :: TRMMDaily{ST,DT}

Creates a TRMMDaily dataset npd to retrieve datasets from the Final post-processing runs for Daily output

Keyword Arguments

start : Date at which download / analysis of the dataset begins

stop : Date at which download / analysis of the dataset ends

path : The directory in which the folder trmmdaily will be created for data downloads, storage and analysis, default is the home directoy called by homedir()

The following fields in npd will be fixed as below:

npdID : trmmdaily

lname : Final TRMM Daily

doi : 10.5067/TRMM/TMPA/DAY/7

hroot : https://disc2.gesdisc.eosdis.nasa.gov/opendap/TRMML3/TRMM3B42_Daily.7

fpref : 3B42_Daily

fsuff : 7.nc4

NASAPrecipitation.TRMMMonthly — Type

TRMMMonthly(
    ST = String,
    DT = Date;
    start :: TimeType,
    stop  :: TimeType,
    path  :: AbstractString = homedir(),
) -> npd :: TRMMMonthly{ST,DT}

Creates a TRMMMonthly dataset npd to retrieve datasets from the Final post-processing runs for Monthly output

Keyword Arguments

start : Date at which download / analysis of the dataset begins

stop : Date at which download / analysis of the dataset ends

path : The directory in which the folder trmmmonthly will be created for data downloads, storage and analysis, default is the home directoy called by homedir()

The following fields in npd will be fixed as below:

npdID : trmmmonthly

lname : TRMM Monthly (TMPA 3B43)

doi : 10.5067/TRMM/TMPA/MONTH/7

hroot : https://disc2.gesdisc.eosdis.nasa.gov/opendap/TRMML3/TRMM3B43.7

fpref : 3B43

fsuff : 7.HDF