Reading the Downloaded NASAPrecipitation Datasets

Once we have downloaded the data, the next step is of course to read and extract the downloaded data. This is done by calling the function read() as follows:

Base.read — Function

read(
    npd :: NASAPrecipitationDataset,
    geo :: GeoRegion,
    dt  :: TimeType;
    lonlat :: Bool = false
) -> NCDataset           (if lonlat = false)
  -> lon, lat, NCDataset (if lonlat = true)

Reads a NASA Precipitation dataset specified by npd for a GeoRegion specified by geo at a date specified by dt.

Arguments

npd : a NASAPrecipitationDataset specifying the dataset details and date download range

geo : a GeoRegion (see GeoRegions.jl) that sets the geographic bounds of the data array in lon-lat

dt : A specified date. The NCDataset retrieved may will contain data for the date, although it may also contain data for other dates depending on the NASAPrecipitationDataset specified by npd

Keyword Arguments

lonlat : if true, then return the longitude and latitude vectors for the dataset. Otherwise only the NCDataset type will be returned.

If for some reason, we need the path to the NetCDF file which contains the data, we can use the function npdfnc() as follows:

NASAPrecipitation.npdfnc — Function

npdfnc(
    npd :: NASAPrecipitationDataset,
    geo :: GeoRegion,
    dt  :: TimeType
) -> String

Returns of the path of the file for the NASA Precipitation dataset specified by npd for a GeoRegion specified by geo at a date specified by dt.

Arguments

npd : a NASAPrecipitationDataset specifying the dataset details and date download range

geo : a GeoRegion (see GeoRegions.jl) that sets the geographic bounds of the data array in lon-lat

dt : A specified date. The NCDataset retrieved may will contain data for the date, although it may also contain data for other dates depending on the NASAPrecipitationDataset specified by npd

using NASAPrecipitation
npd = IMERGMonthly(dtbeg=Date(2015,1,1),dtend=Date(2015,1,1)) # Downloads IMERG Monthly Data for the year 2015
geo = GeoRegion("AR6_SEA")
download(npd,geo)
read(npd,geo,Date(2015,1,1))