Downloading and Saving NASA Precipitation Datasets

In this page we go through downloading NASAPrecipitationDatasets and the Land-Sea Masks for both IMERG and TRMM data.

Required dependencies

Since we are downloading from NASA's EOSDIS OPeNDAP servers, you are required to perform the following:

You need to register an account with Earthdata

Create a .netrc file with the following information: machine urs.earthdata.nasa.gov login <your login> password <your password>

Create a .dodsrc file with the following lines: (1) HTTP.COOKIEJAR=/<home directory>/.urs_cookies and (2) HTTP.NETRC=/<home directory>/.netrc

If this sounds complicated however, fear not! You need only perform the first step yourself (i.e. create your own account), for NASAPrecipitation.jl will automatically set up the .dodsrc file (if you don't already have one), and once you have your <login> and <password>, you can use the function setup() to set up your .netrc file.

setup(
    login = <username>
    password = <password>
)

Downloading NASAPrecipitationDatasets

Downloading datasets from NASA's EOSDIS OPeNDAP servers requires us to specify a NASAPrecipitationDataset npd to download, and a GeoRegion geo to specify the geographic area of interest. This is passed to the download function download(npd,geo).

For a full example on downloading these datasets, look at the page Downloading IMERG Data.

API

NASAPrecipitation.setup — Function

setup(;
    login     :: AbstractString = "",
    password  :: AbstractString = "",
    overwrite :: Bool = false
)

Function that sets up your .dodsrc and .netrc files for you. If you do not provide the keyword arguments for login and password, the .netrc file will not be created.

Keyword arguments

login : Your Earthdata username. You need to specify both login and password arguments.

password : Your Earthdata password. You need to specify both login and password arguments.

overwrite : If true, overwrite existing .dodsrc and .netrc files with the data provided.

Base.download — Function

download(
    npd :: NASAPrecipitationDataset,
    geo :: GeoRegion = GeoRegion("GLB")
) -> nothing

Downloads a NASA Precipitation dataset specified by npd for a GeoRegion specified by geo

Arguments

npd : a NASAPrecipitationDataset specifying the dataset details and date download range

geo : a GeoRegion (see GeoRegions.jl) that sets the geographic bounds of the data array in lon-lat