Extracting data from existing data into a subGeoRegion

One of the drawbacks of retrieving data from the OPeNDAP server is that the connection is slower than using the direct-download method. This cost can become especially noticeable when we are attempting to retrieve data for smaller regions of interest, where the lag caused by opening and closing remote datasets can add up and even take a longer time than doing a direct download.

Thus, drawing upon the functionality of extractGrid() in GeoRegions.jl, we have added in the ability to extract data for a smaller GeoRegion, provided that data for a larger GeoRegion has already been downloaded.

The extraction of data is as easy as

extract(
    <NASAPrecipitation Dataset>,
    <subGeoRegion>,
    <parentGeoRegion>,
)

Setup

using NASAPrecipitation
using CairoMakie
using DelimitedFiles

download("https://raw.githubusercontent.com/natgeo-wong/GeoPlottingData/main/coastline_resl.txt","coast.cst")
coast = readdlm("coast.cst",comments=true)
clon  = coast[:,1]
clat  = coast[:,2]
nothing

An Example: Extracting Data for Southeast Asia from a larger Tropical Domain

Let's define the Tropical Domain and download some NASAPrecipitation data for this region:

npd = TRMMMonthly(start=Date(2015),stop=Date(2015))
pgeo = RectRegion("TRP","GLB","Tropics",[30,-30,360,0],savegeo=false)
download(npd,pgeo)

[ Info: 2023-06-21T16:00:29.101 - NASAPrecipitation.jl - Setting up data structure containing information on TRMM Monthly data to be downloaded
[ Info: 2023-06-21T16:00:29.188 - GeoRegions.jl - Checking to see if the ID TRP is in use
┌ Warning: 2023-06-21T16:00:29.188 - GeoRegions.jl - TRP is not a valid GeoRegion identifier, use either RectRegion() or PolyRegion() to add this GeoRegion to the list.
└ @ GeoRegions ~/.julia/packages/GeoRegions/EQFuG/src/georegions/read.jl:489
[ Info: 2023-06-21T16:00:29.188 - GeoRegions.jl - Adding the GeoRegion TRP to the list.
[ Info: 2023-06-21T16:00:29.195 - GeoRegions.jl - Retrieving information for the GeoRegion defined by the ID GLB
[ Info: 2023-06-21T16:00:29.239 - GeoRegions.jl - Performing a check to determine if the Tropics GeoRegion (TRP) is inside the Globe GeoRegion (GLB)
[ Info: 2023-06-21T16:00:29.239 - GeoRegions.jl - The GeoRegion TRP (Tropics) is indeed a subset of the GeoRegion GLB (Globe)
[ Info: 2023-06-21T16:00:30.509 - NASAPrecipitation.jl - Downloading TRMM Monthly (TMPA 3B43) data for the Tropics GeoRegion from 2015-01-01 to 2015-12-31
[ Info: 2023-06-21T16:00:30.509 - GeoRegions.jl - Retrieving information for the GeoRegion defined by the ID TRMM
[ Info: 2023-06-21T16:00:30.511 - GeoRegions.jl - Performing a check to determine if the Tropics GeoRegion (TRP) is inside the Global TRMM Dataset GeoRegion (TRMM)
[ Info: 2023-06-21T16:00:30.511 - GeoRegions.jl - The GeoRegion TRP (Tropics) is indeed a subset of the GeoRegion TRMM (Global TRMM Dataset)
[ Info: 2023-06-21T16:00:30.511 - GeoRegions.jl - Creating a RegionGrid for the Tropics GeoRegion
[ Info: 2023-06-21T16:00:30.511 - GeoRegions.jl - Creating vector of latitude indices to extract ...
[ Info: 2023-06-21T16:00:30.511 - GeoRegions.jl - Creating vector of longitude indices to extract ...
[ Info: 2023-06-21T16:00:30.511 - NASAPrecipitation.jl - Preallocating temporary arrays for extraction of TRMM Monthly (TMPA 3B43) data for the Tropics GeoRegion from the original gridded dataset
[ Info: 2023-06-21T16:00:30.537 - NASAPrecipitation.jl - Downloading TRMM Monthly (TMPA 3B43) data for the Tropics GeoRegion from the NASA Earthdata servers using OPeNDAP protocols for 2015 ...
[ Info: 2023-06-21T16:01:01.213 - NASAPrecipitation.jl - Saving TRMM Monthly (TMPA 3B43) data in the Tropics GeoRegion for 2015-01-01
[ Info: 2023-06-21T16:01:01.213 - NASAPrecipitation.jl - Creating NetCDF file /home/runner/trmmmonthly/TRP/trmmmonthly-TRP-2015.nc ...
[ Info: 2023-06-21T16:01:01.227 - NASAPrecipitation.jl - TRMM Monthly (TMPA 3B43) data in the Tropics GeoRegion for 2015-01-01 has been saved into /home/runner/trmmmonthly/TRP/trmmmonthly-TRP-2015.nc

We then proceed to define a subGeoRegion, and then extract the data

sgeo = GeoRegion("AR6_SEA")
extract(npd,sgeo,pgeo)

Now, let us proceed to compare and contrast the two Regions

ds = read(npd,pgeo,Date(2015))
prcp_pgeo = ds["precipitation"][:] * 3600
lsd_pgeo  = getLandSea(npd,pgeo)
close(ds)
ds = read(npd,sgeo,Date(2015))
prcp_sgeo = ds["precipitation"][:] * 3600
lsd_sgeo  = getLandSea(npd,sgeo)
close(ds)

fig = Figure()
_,_,slon,slat = coordGeoRegion(sgeo)
aspect = (maximum(slon)-minimum(slon)+90)/(maximum(slat)-minimum(slat)+30)

ax = Axis(
    fig[1,1],width=750,height=750/aspect,
    title="February 2015",xlabel="Longitude / º",ylabel="Latitude / º",
    limits=(minimum(slon)-45,maximum(slon)+45,minimum(slat)-15,maximum(slat)+15)
)
contourf!(
    ax,lsd_pgeo.lon,lsd_pgeo.lat,prcp_pgeo[:,:,2],colormap=:Greens,
    levels=range(0,0.5,length=11),extendlow=:auto,extendhigh=:auto
)
contourf!(
    ax,lsd_sgeo.lon,lsd_sgeo.lat,prcp_sgeo[:,:,2],colormap=:Blues,
    levels=range(0,0.5,length=11),extendlow=:auto,extendhigh=:auto
)
lines!(ax,clon,clat,color=:black)
lines!(ax,slon,slat,color=:red,linewidth=5)

resize_to_layout!(fig)
fig

API

NASAPrecipitation.extract — Function

extract(
    npd :: NASAPrecipitationDataset,
    geo :: GeoRegion
) -> nothing

Extracts NASAPrecipitation data for a GeoRegion geo from its parent GeoRegion geo.pID.

Note

Data for the parent GeoRegion identified by geo.pID must already exist.

Arguments

npd : a NASAPrecipitationDataset specifying the dataset details and date download range

geo : a GeoRegion (see GeoRegions.jl) that sets the geographic bounds of the data array in lon-lat

extract(
    npd  :: NASAPrecipitationDataset,
    sgeo :: GeoRegion,
    pgeo :: GeoRegion,
) -> nothing

Extracts NASAPrecipitation data for a GeoRegion sgeo from a bigger GeoRegion pgeo.

Note

Data for the parent GeoRegion identified by pgeo must already exist, and sgeo must be fully within the bounds of pgeo.

Arguments

npd : a NASAPrecipitationDataset specifying the dataset details and date download range

sgeo : a GeoRegion (see GeoRegions.jl) that sets the geographic bounds of the data array in lon-lat

pgeo : a GeoRegion that is a "parent" of the GeoRegion of interest sgeo, in the sense that sgeo must be fully within pgeo.