NASA Precipitation Datasets

The precipitation datasets that NASAPrecipitation.jl v0.1 currently supports are retrieved from NASA's EOSDIS OPeNDAP servers. They are considered to be a NASAPrecipitationDataset type.

NASAPrecipitationDataset

The NASAPrecipitationDataset AbstractType is meant as an overarching type container for information that identifies various NASA Precipitation Datasets.

NASAPrecipitation.NASAPrecipitationDataset — Type

NASAPrecipitationDataset{ST<:AbstractString, DT<:TimeType}
    npdID :: ST
    lname :: ST
    doi   :: ST
    start :: DT
    stop  :: DT
    datapath :: ST
    maskpath :: ST
    hroot :: ST
    fpref :: ST
    fsuff :: ST
end

Abstract supertype for NASA Precipitation datasets on NASA OPeNDAP Servers.

Fields:

npdID : ID for the NASAPrecipitationDataset, used in determining containing folders and filenames of the NetCDF

lname : The name describing the NASAPrecipitationDataset, used mostly in Logging

doi : The DOI identifier, to be saved into the NetCDF

start : The start date (Y,M,D) of our download / analysis

stop : The end date (Y,M,D) of our download / analysis

datapath : The directory in which to save our downloads and analysis files to

maskpath : The directory in which to save our corresponding landsea mask datasets

hroot : The URL of the NASA's EOSDIS OPeNDAP server for which this dataset is stored

fpref : The prefix component of the NetCDF files to be downloaded

fsuff : The suffix component of the NetCDF files to be downloaded

Of these fields, only start, stop, datapath and maskpath are user-defined. All other fields are predetermined depending on the type of NASA Precipitation Dataset called.

All subTypes of the NASAPrecipitationDataset have the same basic fields and structure.

julia> using NASAPrecipitation
julia> npd = IMERGFinalHH(dtbeg=Date(2017,2,1),dtend=Date(2017,2,1),sroot=homedir())ERROR: MethodError: no method matching IMERGFinalHH(::Type{String}, ::Type{Date}; dtbeg=Date("2017-02-01"), dtend=Date("2017-02-01"), sroot="/home/runner")
Closest candidates are:
  IMERGFinalHH(::Any, ::Any; start, stop, path) at ~/work/NASAPrecipitation.jl/NASAPrecipitation.jl/src/datasets/imerghh.jl:144 got unsupported keyword arguments "dtbeg", "dtend", "sroot"
  IMERGFinalHH(::Any) at ~/work/NASAPrecipitation.jl/NASAPrecipitation.jl/src/datasets/imerghh.jl:144 got unsupported keyword arguments "dtbeg", "dtend", "sroot"
julia> npd.dtbegERROR: UndefVarError: npd not defined
julia> npd.srootERROR: UndefVarError: npd not defined
julia> npd.fprefERROR: UndefVarError: npd not defined
julia> typeof(npd)ERROR: UndefVarError: npd not defined
julia> typeof(npd) <: NASAPrecipitation.NASAPrecipitationDatasetERROR: UndefVarError: npd not defined

Supported NASA Datasets

The following datasets are supported in v0.1:

GPM IMERGv6, 0.1º resolution, 60ºS to 60ºN, represented by the IMERGDataset Type

Final runs for Half-Hourly, Daily and Monthly Data

Early and Late runs of Half-Hourly and Daily Data

TRMM TMPAv7, 0.25º resolution, 50ºS to 50ºN, represented by the TRMMataset Type

Final runs for 3-Hourly, Daily and Monthly Data

Near Real-Time runs of 3-Hourly and Daily Data

NASAPrecipitation.IMERGDataset — Type

IMERGDataset <: NASAPrecipitationDataset

Abstract supertype for GPM IMERG datasets on NASA OPeNDAP Servers, a subType of NASAPrecipitationDataset.

NASAPrecipitation.TRMMDataset — Type

TRMMDataset <: NASAPrecipitationDataset

Abstract supertype for TRMM TMPA datasets on NASA OPeNDAP Servers, a subType of NASAPrecipitationDataset.

Each of these Abstract superTypes has its own set of Types. See the respective pages for GPM IMERG and TRMM TMPA for more information.