The API for NASAPrecipitation.jl Downloads

In this page we go through downloading NASAPrecipitationDatasets and the Land-Sea Masks for both IMERG and TRMM data.

Required dependencies

Since we are downloading from NASA's EOSDIS OPeNDAP servers, you need to do the following steps before NASAPrecipitation downloads will work:

You need to register an account with Earthdata

Create a .netrc file, and paste the following:

machine urs.earthdata.nasa.gov login <your login> password <your password>

Create a .dodsrc file, and paste the following:

HTTP.COOKIEJAR=/<home directory>/.urs_cookies

HTTP.NETRC=/<home directory>/.netrc

Downloading NASAPrecipitationDatasets

Downloading datasets from NASA's EOSDIS OPeNDAP servers requires us to specify a NASAPrecipitationDataset npd to download, and a GeoRegion geo to specify the geographic area of interest. This is passed to the download function download(npd,geo).

For a full example on downloading these datasets, look at the page Downloading IMERG Data.

Base.download — Function

download(
    npd :: NASAPrecipitationDataset,
    geo :: GeoRegion = GeoRegion("GLB")
) -> nothing

Downloads a NASA Precipitation dataset specified by npd for a GeoRegion specified by geo

Arguments

npd : a NASAPrecipitationDataset specifying the dataset details and date download range

geo : a GeoRegion (see GeoRegions.jl) that sets the geographic bounds of the data array in lon-lat

For a basic introduction to the concept of a GeoRegion, you can refer to the page A Basic Primer to GeoRegions. Alternatively you can refer to the GeoRegions.jl documentation for a comprehensive tutorial.

Downloading Land-Sea Masks

We are also able to use NASAPrecipitation.jl to download Land-Sea Masks on the same grids as IMERGDatasets and TRMMDatasets. This is of great help in allowing us to filter data based on whether it is over land or sea.

For a full example on how these land-sea masks can be used in dataset analysis, look at the page Land-Sea Mask Filtering.

Missing docstring.

Missing docstring for NASAPrecipitation.getIMERGlsm. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NASAPrecipitation.getTRMMlsm. Check Documenter's build log for details.